agent/requirements.txt
livekit-agents>=0.12.0
livekit-plugins-silero>=0.7.0
faster-whisper>=1.0.0
httpx>=0.27.0
redis>=5.0.0
